--- Maciek OS Core.sln
@@ -1,42 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.1.31903.286
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Maciek OS Core", "Maciek OS Core\Maciek OS Core.csproj", "{1E93AC15-A78D-4D51-BE79-8A34AEB5420B}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{AAE1FA50-2A7D-4E01-85F1-8D3AC9275F1B}"
-	ProjectSection(SolutionItems) = preProject
-		.editorconfig = .editorconfig
-	EndProjectSection
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MOS User Menager Integration", "..\MOS User Menager Integration\MOS User Menager Integration\MOS User Menager Integration.csproj", "{9CD6F790-D5F2-4527-89B6-3AB8D5C99D48}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MOS Log Integration", "..\MOS Log Integration\MOS Log Integration\MOS Log Integration.csproj", "{6E8405C5-91E3-47F2-AD8B-449932542AFE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1E93AC15-A78D-4D51-BE79-8A34AEB5420B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1E93AC15-A78D-4D51-BE79-8A34AEB5420B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1E93AC15-A78D-4D51-BE79-8A34AEB5420B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1E93AC15-A78D-4D51-BE79-8A34AEB5420B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9CD6F790-D5F2-4527-89B6-3AB8D5C99D48}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9CD6F790-D5F2-4527-89B6-3AB8D5C99D48}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9CD6F790-D5F2-4527-89B6-3AB8D5C99D48}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9CD6F790-D5F2-4527-89B6-3AB8D5C99D48}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6E8405C5-91E3-47F2-AD8B-449932542AFE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6E8405C5-91E3-47F2-AD8B-449932542AFE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6E8405C5-91E3-47F2-AD8B-449932542AFE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6E8405C5-91E3-47F2-AD8B-449932542AFE}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {AE39D1FC-C3CF-41A9-873A-BBF37F8C9ED0}
-	EndGlobalSection
-EndGlobal
