--- Maciek OS Core/App.config
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-    <configSections>
-        <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
-            <section name="Maciek_OS_Core.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false"/>
-        </sectionGroup>
-    </configSections>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.8"/>
-    </startup>
-    <userSettings>
-        <Maciek_OS_Core.Properties.Settings>
-            <setting name="Experimental" serializeAs="String">
-                <value>True</value>
-            </setting>
-            <setting name="Version" serializeAs="String">
-                <value>4.3</value>
-            </setting>
-            <setting name="Compiled" serializeAs="String">
-                <value>03.02</value>
-            </setting>
-            <setting name="year" serializeAs="String">
-                <value>2022</value>
-            </setting>
-            <setting name="Build" serializeAs="String">
-                <value>0275</value>
-            </setting>
-        </Maciek_OS_Core.Properties.Settings>
-    </userSettings>
-</configuration>
