--- Maciek OS Core/Commands/Cmds/Nano/Text.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-namespace Maciek_OS_Core.Commands.Cmds.Nano
-{
-	public class Text
-	{
-		public int N;
-		public string S;
-		public ConsoleColor Color = ConsoleColor.White;
-		public ConsoleColor BGColor = ConsoleColor.Black;
-		public Text(int n, string s)
-		{
-			N = n;
-			S = s;
-		}
-	}
-}
