+++ Maciek SHELL/GlobalSuppressions.cs
@@ -1,0 +1,9 @@
+﻿// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("CodeQuality", "IDE0052:Usuń nieodczytywane składowe prywatne", Justification = "Critical variable", Scope = "member", Target = "~F:Maciek_SHELL.Program.userController")]
+[assembly: SuppressMessage("CodeQuality", "IDE0052:Usuń nieodczytywane składowe prywatne", Justification = "Critical variable", Scope = "member", Target = "~F:Maciek_SHELL.Program.log")]
