+++ Maciek SHELL/Properties/AssemblyInfo.cs
