+++ Maciek SHELL/Properties/Settings.settings
@@ -1,0 +1,15 @@
+﻿<?xml version='1.0' encoding='utf-8'?>
+<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="Maciek_SHELL.Properties" GeneratedClassName="Settings">
+  <Profiles />
+  <Settings>
+    <Setting Name="Experimental" Type="System.Boolean" Scope="User">
+      <Value Profile="(Default)">True</Value>
+    </Setting>
+    <Setting Name="Version" Type="System.String" Scope="User">
+      <Value Profile="(Default)">4.3</Value>
+    </Setting>
+    <Setting Name="Build" Type="System.String" Scope="User">
+      <Value Profile="(Default)">0275</Value>
+    </Setting>
+  </Settings>
+</SettingsFile>
