+++ Maciek SHELL/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="SimpleLogs4Net" version="1.2.0" targetFramework="net48" />
+</packages>
